+++ packages/plugins/ai/tsup.config.ts
@@ -3,6 +3,6 @@ import { sharedConfig } from '../tsup.shared.js';
 
 export default defineConfig({
 	...sharedConfig,
-	entry: ['src/index.ts'],
+	entry: ['src/index.ts', 'src/install.ts'],
 	dts: true,
 });