+++ packages/plugins/batch/tsup.config.ts
@@ -4,5 +4,5 @@ import { sharedConfig } from '../tsup.shared.js';
 export default defineConfig({
 	...sharedConfig,
 	format: ['esm'], // Node.js specific plugin
-	entry: ['src/index.ts', 'src/worker.ts'],
+	entry: ['src/index.ts', 'src/install.ts', 'src/worker.ts'],
 });