+++ packages/plugins/celestial/tsup.config.ts
@@ -3,5 +3,5 @@ import { sharedConfig } from '../tsup.shared.js';
 
 export default defineConfig({
 	...sharedConfig,
-	entry: ['src/index.ts'],
+	entry: ['src/index.ts', 'src/install.ts'],
 });