+++ packages/plugins/snap/src/install.ts
@@ -1,0 +1,6 @@
+import { autoInstall } from '@magmacomputing/tempo/plugin/sdk';
+import { SnapPlugin } from './index.js';
+
+autoInstall(SnapPlugin);
+
+export * from './index.js';